Change packages names, add service package and add product service
