+++ src/main/java/com/ecommerce/games/games/entities/Rating.java
@@ -1,4 +1,4 @@
-package com.ecommerce.games.games.entity;
+package com.ecommerce.games.games.entities;
 
 public enum Rating {
     EVERYONE(1),
