+++ src/main/java/com/ecommerce/games/games/repositories/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.ecommerce.games.games.repository;
+package com.ecommerce.games.games.repositories;
 
-import com.ecommerce.games.games.entity.Category;
+import com.ecommerce.games.games.entities.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
