+++ src/main/java/com/ecommerce/games/games/repositories/ProductRepository.java
@@ -1,6 +1,6 @@
-package com.ecommerce.games.games.repository;
+package com.ecommerce.games.games.repositories;
 
-import com.ecommerce.games.games.entity.Product;
+import com.ecommerce.games.games.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
