+++ src/main/resources/application.properties
@@ -6,5 +6,5 @@ spring.datasource.password=
 spring.datasource.driver-class-name=org.h2.Driver
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.hibernate.ddl-auto=none
+spring.jpa.hibernate.ddl-auto=create
 # No JDBC URL: jdbc:h2:file:~/test
